code/rpi/main/requirements.txt
schedule
RPi.GPIO
board
neopixel
picamera
prometheus_client
